Fix sponsor links.
The topmost element is still technically visible even when its opacity
is 0, so its link will take precedent. Whoops! Better change visibility
along with opacity, then. We change visibility, not display, so that the
footprint of the elements stays constant (i.e., so that we don't change
the page height as logos cycle).